fix(module7-task3): Добавил компонент RatingInput
