--- project/src/components/form.tsx
@@ -1,155 +1,0 @@
-import { useState } from 'react';
-import { useAppDispatch } from '../hooks';
-import { useParams } from 'react-router-dom';
-import { postComment } from '../store/api-actions';
-
-
-export const Form = () => {
-  const { id = '' } = useParams();
-  const dispatch = useAppDispatch();
-  const [state, setState] = useState({
-    rating: 0,
-    comment: '',
-  });
-
-  const handleRatingChange: React.ChangeEventHandler<HTMLInputElement> = (event) => {
-    const { value } = event.target;
-    setState({
-      ...state,
-      rating: Number(value)
-    });
-  };
-
-  const handleTextChange: React.ChangeEventHandler<HTMLTextAreaElement> = (event) => {
-    const { value } = event.target;
-    setState({
-      ...state,
-      comment: value
-    });
-  };
-
-  const handleSubmit = () => { dispatch(postComment({ ...state, id })); };
-
-  return (
-    <form className='reviews__form form' onSubmit={handleSubmit}>
-      <label className='reviews__label form__label' htmlFor='review'>
-        Your review
-      </label>
-      <div className='reviews__rating-form form__rating'>
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          value={5}
-          id='5-stars'
-          type='radio'
-          onChange={handleRatingChange}
-        />
-        <label
-          htmlFor='5-stars'
-          className='reviews__rating-label form__rating-label'
-          title='perfect'
-        >
-          <svg className='form__star-image' width='37' height='33'>
-            <use xlinkHref='#icon-star'></use>
-          </svg>
-        </label>
-
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          value={4}
-          id='4-stars'
-          type='radio'
-          onChange={handleRatingChange}
-        />
-        <label
-          htmlFor='4-stars'
-          className='reviews__rating-label form__rating-label'
-          title='good'
-        >
-          <svg className='form__star-image' width='37' height='33'>
-            <use xlinkHref='#icon-star'></use>
-          </svg>
-        </label>
-
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          value={3}
-          id='3-stars'
-          type='radio'
-          onChange={handleRatingChange}
-        />
-        <label
-          htmlFor='3-stars'
-          className='reviews__rating-label form__rating-label'
-          title='not bad'
-        >
-          <svg className='form__star-image' width='37' height='33'>
-            <use xlinkHref='#icon-star'></use>
-          </svg>
-        </label>
-
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          value={2}
-          id='2-stars'
-          type='radio'
-          onChange={handleRatingChange}
-        />
-        <label
-          htmlFor='2-stars'
-          className='reviews__rating-label form__rating-label'
-          title='badly'
-        >
-          <svg className='form__star-image' width='37' height='33'>
-            <use xlinkHref='#icon-star'></use>
-          </svg>
-        </label>
-
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          value={1}
-          id='1-star'
-          type='radio'
-          onChange={handleRatingChange}
-        />
-        <label
-          htmlFor='1-star'
-          className='reviews__rating-label form__rating-label'
-          title='terribly'
-        >
-          <svg className='form__star-image' width='37' height='33'>
-            <use xlinkHref='#icon-star'></use>
-          </svg>
-        </label>
-      </div>
-      <textarea
-        className='reviews__textarea form__textarea'
-        id='review'
-        name='review'
-        value={state.comment}
-        placeholder='Tell how was your stay, what you like and what can be improved'
-        onChange={handleTextChange}
-      />
-      <div className='reviews__button-wrapper'>
-        <p className='reviews__help'>
-          To submit review please make sure to set
-          <span className='reviews__star'>rating</span> and describe your stay
-          with at least
-          <b className='reviews__text-amount'>50 characters</b>.
-        </p>
-        <button
-          className='reviews__submit form__submit button'
-          type='submit'
-          disabled={state.comment.length < 50 || state.comment.length > 250}
-          title="Length of message should be less than 250 characters"
-        >
-          Submit
-        </button>
-      </div>
-    </form>
-  );
-};
